@@ -75,14 +75,67 @@
     "logo": "https://opencollective.com/opencollective/logo.txt"
   },
   "devDependencies": {
-    "awesome-typescript-loader": "^3.4.1",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.0.3",
+    "babel-plugin-import": "^1.7.0",
+    "babel-plugin-styled-components": "^1.3.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.12",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-flow": "^6.23.0",
+    "babel-preset-minify": "^0.5.0-alpha.a24dd066",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "cache-loader": "^1.2.2",
+    "css-loader": "^1.0.0",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-import-resolver-webpack": "^0.9.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-json": "^1.2.0",
+    "eslint-plugin-jsx-a11y": "6.0.3",
+    "eslint-plugin-mocha": "^5.0.0",
+    "eslint-plugin-prettier": "^2.4.0",
+    "eslint-plugin-react": "^7.5.1",
+    "file-loader": "^1.1.6",
+    "fork-ts-checker-webpack-plugin": "^0.4.9",
+    "heroku-babel-loader": "^7.1.2",
+    "html-webpack-plugin": "^3.2.0",
+    "http-proxy-middleware": "^0.18.0",
     "husky": "^0.14.3",
+    "ignore-loader": "^0.1.2",
+    "isomorphic-style-loader": "^4.0.0",
     "lerna": "^2.5.1",
+    "less": "^3.0.2",
+    "less-loader": "^4.1.0",
     "lint-staged": "^7.1.2",
-    "tslint": "^5.2.0",
+    "mini-css-extract-plugin": "^0.4.1",
+    "node-sass": "^4.7.2",
+    "openurl": "^1.1.1",
+    "postcss-loader": "^3.0.0",
+    "prettier": "^1.9.2",
+    "qrcode-terminal": "^0.12.0",
+    "raw-loader": "^0.5.1",
+    "sass-loader": "^7.1.0",
+    "source-list-map": "^2.0.0",
+    "style-loader": "^0.21.0",
+    "ts-loader": "^4.5.0",
+    "tslint": "^5.11.0",
     "tslint-config-prettier": "^1.5.0",
     "tslint-plugin-prettier": "^1.3.0",
-    "typescript": "^2.7.2"
+    "typescript": "^2.9.2",
+    "url-loader": "^1.0.1",
+    "wait-on": "^2.0.2",
+    "webpack": "^4.17.0",
+    "webpack-dev-middleware": "^3.1.3",
+    "webpack-dev-server": "^3.1.5",
+    "webpack-hot-middleware": "^2.22.3",
+    "webpack-manifest-plugin": "^2.0.3",
+    "webpack-merge": "^4.1.4",
+    "webpack-node-externals": "^1.7.2",
+    "webpack-sources": "^1.1.0",
+    "webpack-virtual-modules": "^0.1.10"
   },
   "optionalDependencies": {
     "dotenv": "^6.0.0",
@@ -90,10 +143,10 @@
   },
   "dependencies": {
     "opencollective": "^1.0.3",
-    "spinjs": "^0.4.145"
+    "spinjs": "^0.4.155"
   },
   "resolutions": {
-    "graphql": "0.13.0",
+    "graphql": "0.13.2",
     "iterall": "^1.2.2",
     "upath": "^1.1.0"
   }
