@@ -13,7 +13,7 @@ class AwakeInDevApp extends React.Component {
     isReady: false
   };
 
-  async UNSAFE_componentWillMount() {
+  async componentDidMount() {
     await Expo.Font.loadAsync({
       Roboto: require('native-base/Fonts/Roboto.ttf'),
       Roboto_medium: require('native-base/Fonts/Roboto_medium.ttf'),
