@@ -1,9 +1,6 @@
 export default {
-  engineConfig: {
-    apiKey: '', // Set your Engine API key here
-    logging: {
-      level: 'DEBUG' // Engine Proxy logging level. DEBUG, INFO, WARN or ERROR
-    }
-  },
-  dumpTraffic: true // Debug configuration that logs traffic between Proxy and GraphQL server
+  apiKey: '', // Set your Engine API key here
+  logging: {
+    level: 'DEBUG' // Engine Proxy logging level. DEBUG, INFO, WARN or ERROR
+  }
 };
