@@ -4,23 +4,26 @@ import { Button as NBButton, Text } from 'native-base';
 import { StyleSheet } from 'react-native';
 import ButtonStyles from '../styles/Button';
 
-const Button = ({ textStyle, children, onClick, onPress, type, size, ...props }) => {
-  const btnProps = {
-    ...props,
-    [type]: true,
-    [size]: true,
-    block: true,
-    onPress: onPress || onClick
-  };
+class Button extends React.Component {
+  render() {
+    const { textStyle, children, onClick, onPress, type, size, ...props } = this.props;
+    const btnProps = {
+      ...props,
+      [type]: true,
+      [size]: true,
+      block: true,
+      onPress: onPress || onClick
+    };
 
-  return (
-    <NBButton {...btnProps}>
-      <Text style={[styles.btnText, textStyle]} numberOfLines={1}>
-        {children}
-      </Text>
-    </NBButton>
-  );
-};
+    return (
+      <NBButton {...btnProps}>
+        <Text style={[styles.btnText, textStyle]} numberOfLines={1}>
+          {children}
+        </Text>
+      </NBButton>
+    );
+  }
+}
 
 const styles = StyleSheet.create(ButtonStyles);
 
