@@ -3,15 +3,18 @@ import PropTypes from 'prop-types';
 import { Button, Icon } from 'native-base';
 import { Feather } from '@expo/vector-icons';
 
-const IconButton = ({ iconName, style, iconColor, iconSize, ...props }) => {
-  return (
-    <Button transparent style={[{ alignSelf: 'center' }, style]} {...props}>
-      <Icon>
-        <Feather name={iconName} size={iconSize} color={iconColor} />
-      </Icon>
-    </Button>
-  );
-};
+class IconButton extends React.Component {
+  render() {
+    const { iconName, style, iconColor, iconSize, ...props } = this.props;
+    return (
+      <Button transparent style={[{ alignSelf: 'center' }, style]} {...props}>
+        <Icon>
+          <Feather name={iconName} size={iconSize} color={iconColor} />
+        </Icon>
+      </Button>
+    );
+  }
+}
 
 IconButton.propTypes = {
   iconName: PropTypes.string,
