@@ -14,21 +14,22 @@ const Footer = styled.footer`
   height: ${footerHeight};
 `;
 
-const PageLayout = ({ children, navBar }) => {
-  return (
-    <section className="d-flex flex-column flex-grow-1">
-      <section className="d-flex flex-column flex-grow-1 flex-shrink-0">
-        <section className="d-flex flex-column">{navBar !== false && <NavBar />}</section>
-        <Container id="content">{children}</Container>
+class PageLayout extends React.Component {
+  render() {
+    const { children, navBar } = this.props;
+    return (
+      <section className="d-flex flex-column flex-grow-1">
+        <section className="d-flex flex-column flex-grow-1 flex-shrink-0">
+          <section className="d-flex flex-column">{navBar !== false && <NavBar />}</section>
+          <Container id="content">{children}</Container>
+        </section>
+        <Footer className="d-flex flex-shrink-0 justify-content-center">
+          <span>&copy; 2017. {settings.app.name}.</span>
+        </Footer>
       </section>
-      <Footer className="d-flex flex-shrink-0 justify-content-center">
-        {/*
-        <span>&copy; 2017. {settings.app.name}.</span>
-        */}
-      </Footer>
-    </section>
-  );
-};
+    );
+  }
+}
 
 PageLayout.propTypes = {
   children: PropTypes.node,
