@@ -2,7 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withFormik } from 'formik';
 
-import translate from '../../../i18n';
 import Field from '../../../utils/FieldAdapter';
 import { Form, RenderField, Button, Alert } from '../../common/components/web';
 import { email, minLength, required, validateForm } from '../../../../../common/validation';
@@ -13,38 +12,32 @@ const contactFormSchema = {
   content: [required, minLength(10)]
 };
 
-const validate = values => validateForm(values, contactFormSchema);
-
-const ContactForm = ({ values, handleSubmit, error, sent, t }) => {
-  return (
-    <Form name="contact" onSubmit={handleSubmit}>
-      {sent && <Alert color="success">{t('form.submitMsg')}</Alert>}
-      <Field name="name" component={RenderField} type="text" label={t('form.field.name')} value={values.name} />
-      <Field name="email" component={RenderField} type="text" label={t('form.field.email')} value={values.email} />
-      <Field
-        name="content"
-        component={RenderField}
-        type="textarea"
-        label={t('form.field.content')}
-        value={values.content}
-      />
-      <div className="text-center">
-        {error && <Alert color="error">{error}</Alert>}
-        <Button color="primary" type="submit">
-          {t('form.btnSubmit')}
-        </Button>
-      </div>
-    </Form>
-  );
-};
+const ContactForm = ({ values, handleSubmit, errors, t, status }) => (
+  <Form name="contact" onSubmit={handleSubmit}>
+    {status && status.sent && <Alert color="success">{t('successMsg')}</Alert>}
+    <Field name="name" component={RenderField} type="text" label={t('form.field.name')} value={values.name} />
+    <Field name="email" component={RenderField} type="text" label={t('form.field.email')} value={values.email} />
+    <Field
+      name="content"
+      component={RenderField}
+      type="textarea"
+      label={t('form.field.content')}
+      value={values.content}
+    />
+    <div className="text-center">
+      {errors._error && <Alert color="error">{errors._error}</Alert>}
+      <Button color="primary" type="submit">
+        {t('form.btnSubmit')}
+      </Button>
+    </div>
+  </Form>
+);
 
 ContactForm.propTypes = {
-  handleSubmit: PropTypes.func,
-  onSubmit: PropTypes.func,
-  submitting: PropTypes.bool,
-  error: PropTypes.string,
-  sent: PropTypes.bool,
   values: PropTypes.object,
+  handleSubmit: PropTypes.func,
+  errors: PropTypes.object,
+  status: PropTypes.object,
   t: PropTypes.func
 };
 
@@ -55,14 +48,22 @@ const ContactFormWithFormik = withFormik({
     values,
     {
       resetForm,
+      setErrors,
+      setStatus,
       props: { onSubmit }
     }
   ) {
-    await onSubmit(values);
-    resetForm();
+    try {
+      await onSubmit(values);
+      resetForm();
+      setStatus({ sent: true });
+    } catch (e) {
+      setStatus({ sent: false });
+      setErrors(e);
+    }
   },
-  validate: values => validate(values),
+  validate: values => validateForm(values, contactFormSchema),
   displayName: 'ContactUsForm' // helps with React DevTools
 });
 
-export default translate('contact')(ContactFormWithFormik(ContactForm));
+export default ContactFormWithFormik(ContactForm);
