@@ -1,4 +1,5 @@
-/* 
+/*
 * The index.js can be empty, it's just needed to point the loader to the root directory of the locales.
 * https://github.com/alienfast/i18next-loader#option-2-use-with-import-syntax
 */
+export default {};
