@@ -3,25 +3,12 @@ import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 import Dropzone from 'react-dropzone';
 import filesize from 'filesize';
-import translate from '../../../i18n';
 
 import { PageLayout, Row, Col, Table, Button, Alert } from '../../common/components/web';
 import settings from '../../../../../../settings';
 
-class UploadView extends React.PureComponent {
-  static propTypes = {
-    files: PropTypes.array,
-    uploadFiles: PropTypes.func.isRequired,
-    removeFile: PropTypes.func.isRequired,
-    t: PropTypes.func
-  };
-
-  state = {
-    error: null
-  };
-
-  renderMetaData = () => {
-    const { t } = this.props;
+const UploadView = ({ files, error, loading, handleUploadFiles, handleRemoveFile, t }) => {
+  const renderMetaData = () => {
     return (
       <Helmet
         title={`${settings.app.name} - ${t('title')}`}
@@ -35,71 +22,61 @@ class UploadView extends React.PureComponent {
     );
   };
 
-  onDrop = uploadFiles => async files => {
-    const result = await uploadFiles(files);
-    if (result && result.error) {
-      this.setState({ error: result.error });
-    } else {
-      this.setState({ error: null });
-    }
-  };
-
-  handleRemoveFile = async id => {
-    const { removeFile } = this.props;
-    const result = await removeFile(id);
-    if (result && result.error) {
-      this.setState({ error: result.error });
-    } else {
-      this.setState({ error: null });
-    }
-  };
-
-  render() {
-    const { files, uploadFiles, t } = this.props;
-    const { error } = this.state;
-
-    const columns = [
-      {
-        title: t('table.column.name'),
-        dataIndex: 'name',
-        key: 'name',
-        render: (text, record) => (
+  const columns = [
+    {
+      title: t('table.column.name'),
+      dataIndex: 'name',
+      key: 'name',
+      render(text, record) {
+        return (
           <a href={record.path} download={text}>
             {text} ({filesize(record.size)})
           </a>
-        )
-      },
-      {
-        title: t('table.column.actions'),
-        key: 'actions',
-        width: 50,
-        render: (text, record) => (
-          <Button color="primary" size="sm" className="delete-button" onClick={() => this.handleRemoveFile(record.id)}>
+        );
+      }
+    },
+    {
+      title: t('table.column.actions'),
+      key: 'actions',
+      width: 50,
+      render(text, record) {
+        return (
+          <Button color="primary" size="sm" className="delete-button" onClick={() => handleRemoveFile(record.id)}>
             {t('table.btnDel')}
           </Button>
-        )
+        );
       }
-    ];
+    }
+  ];
 
-    return (
-      <PageLayout>
-        {this.renderMetaData()}
-        <div className="text-center">
-          <Row>
-            <Col xs={4}>
-              <Dropzone onDrop={this.onDrop(uploadFiles)}>
-                <p>{t('message')}</p>
-              </Dropzone>
-            </Col>
-            <Col xs={8}>
-              {error && <Alert color="error">{error}</Alert>}
-              {files && <Table dataSource={files} columns={columns} />}
-            </Col>
-          </Row>
-        </div>
-      </PageLayout>
-    );
-  }
-}
+  return (
+    <PageLayout>
+      {renderMetaData()}
+      <div className="text-center">
+        <Row>
+          <Col xs={4}>
+            <Dropzone onDrop={handleUploadFiles}>
+              <p>{t('message')}</p>
+            </Dropzone>
+          </Col>
+          <Col xs={8}>
+            {loading && <span>Loading...</span>}
+            {error && <Alert color="error">{error}</Alert>}
+            {files && <Table dataSource={files} columns={columns} />}
+          </Col>
+        </Row>
+      </div>
+    </PageLayout>
+  );
+};
 
-export default translate('upload')(UploadView);
+UploadView.propTypes = {
+  files: PropTypes.array,
+  error: PropTypes.string,
+  loading: PropTypes.bool,
+  handleUploadFiles: PropTypes.func.isRequired,
+  handleRemoveFile: PropTypes.func.isRequired,
+  t: PropTypes.func
+};
+
+export default UploadView;
