@@ -5,36 +5,63 @@ import UploadView from '../components/UploadView';
 import FILES_QUERY from '../graphql/FilesQuery.graphql';
 import UPLOAD_FILES from '../graphql/UploadFiles.graphql';
 import REMOVE_FILE from '../graphql/RemoveFile.graphql';
+import translate from '../../../i18n';
 
 class Upload extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      error: null
+    };
+  }
+
+  handleUploadFiles = async files => {
+    const { uploadFiles } = this.props;
+    const result = await uploadFiles(files);
+
+    this.setState({ error: result && result.error ? result.error : null });
+  };
+
+  handleRemoveFile = async id => {
+    const { removeFile } = this.props;
+    const result = await removeFile(id);
+
+    this.setState({ error: result && result.error ? result.error : null });
+  };
+
   render() {
-    return <UploadView {...this.props} />;
+    return (
+      <UploadView
+        {...this.props}
+        error={this.state.error}
+        handleRemoveFile={this.handleRemoveFile}
+        handleUploadFiles={this.handleUploadFiles}
+      />
+    );
   }
 }
 
-const UploadWithApollo = compose(
+export default compose(
   graphql(FILES_QUERY, {
     options: () => {
       return {
         fetchPolicy: 'cache-and-network'
       };
     },
-    props({ data: { loading, error, files, refetch } }) {
+    props({ data: { loading, error, files } }) {
       if (error) throw new Error(error);
-      return { loading, files, refetch };
+
+      return { loading, files };
     }
   }),
   graphql(UPLOAD_FILES, {
-    props: ({ ownProps: { refetch }, mutate }) => ({
+    props: ({ mutate }) => ({
       uploadFiles: async files => {
         try {
-          const {
-            data: { uploadFiles }
-          } = await mutate({
-            variables: { files }
+          await mutate({
+            variables: { files },
+            refetchQueries: [{ query: FILES_QUERY }]
           });
-          refetch();
-          return uploadFiles;
         } catch (e) {
           return { error: e.graphQLErrors[0].message };
         }
@@ -42,22 +69,32 @@ const UploadWithApollo = compose(
     })
   }),
   graphql(REMOVE_FILE, {
-    props: ({ ownProps: { refetch }, mutate }) => ({
+    props: ({ mutate }) => ({
       removeFile: async id => {
         try {
-          const {
-            data: { removeFile }
-          } = await mutate({
-            variables: { id }
+          await mutate({
+            variables: { id },
+            optimisticResponse: {
+              __typename: 'Mutation',
+              removeFile: {
+                removeFile: true,
+                __typename: 'File'
+              }
+            },
+            update: store => {
+              const cachedFiles = store.readQuery({ query: FILES_QUERY });
+
+              store.writeQuery({
+                query: FILES_QUERY,
+                data: { files: cachedFiles.files.filter(file => file.id !== id) }
+              });
+            }
           });
-          refetch();
-          return removeFile;
         } catch (e) {
           return { error: e.graphQLErrors[0].message };
         }
       }
     })
-  })
+  }),
+  translate('upload')
 )(Upload);
-
-export default UploadWithApollo;
