@@ -1,5 +1,6 @@
 import { graphql } from 'react-apollo';
 import update from 'immutability-helper';
+
 import { removeTypename } from '../../../../../common/utils';
 
 import USERS_STATE_QUERY from '../graphql/UsersStateQuery.client.graphql';
@@ -7,7 +8,6 @@ import UPDATE_ORDER_BY from '../graphql/UpdateOrderBy.client.graphql';
 import USERS_QUERY from '../graphql/UsersQuery.graphql';
 import DELETE_USER from '../graphql/DeleteUser.graphql';
 import UPDATE_FILTER from '../graphql/UpdateFilter.client.graphql';
-import USERS_SUBSCRIPTION from '../graphql/UsersSubscription.graphql';
 
 const withUsersState = Component =>
   graphql(USERS_STATE_QUERY, {
@@ -24,8 +24,8 @@ const withUsers = Component =>
         variables: { orderBy, filter }
       };
     },
-    props({ data: { loading, users, refetch, error, subscribeToMore } }) {
-      return { loading, users, refetch, subscribeToMore, errors: error ? error.graphQLErrors : null };
+    props({ data: { loading, users, refetch, error, updateQuery } }) {
+      return { loading, users, refetch, updateQuery, errors: error ? error.graphQLErrors : null };
     }
   })(Component);
 
@@ -74,7 +74,28 @@ const withFilterUpdating = Component =>
     })
   })(Component);
 
+const updateUsersState = (usersUpdated, updateQuery) => {
+  const { mutation, node } = usersUpdated;
+  updateQuery(prev => {
+    switch (mutation) {
+      case 'CREATED':
+        return addUser(prev, node);
+      case 'DELETED':
+        return deleteUser(prev, node.id);
+      case 'UPDATED':
+        return deleteUser(prev, node.id);
+      default:
+        return prev;
+    }
+  });
+};
+
 function addUser(prev, node) {
+  // check if it is duplicate
+  if (prev.users.some(user => user.id === node.id)) {
+    return prev;
+  }
+
   return update(prev, {
     users: {
       $set: [...prev.users, node]
@@ -95,33 +116,5 @@ function deleteUser(prev, id) {
   });
 }
 
-const subscribeToUsersList = (subscribeToMore, filter) => {
-  return subscribeToMore({
-    document: USERS_SUBSCRIPTION,
-    variables: { filter },
-    updateQuery: (
-      prev,
-      {
-        subscriptionData: {
-          data: {
-            usersUpdated: { mutation, node }
-          }
-        }
-      }
-    ) => {
-      switch (mutation) {
-        case 'CREATED':
-          return addUser(prev, node);
-        case 'DELETED':
-          return deleteUser(prev, node.id);
-        case 'UPDATED':
-          return deleteUser(prev, node.id);
-        default:
-          return prev;
-      }
-    }
-  });
-};
-
 export { withUsersState, withUsers, withUsersDeleting, withOrderByUpdating, withFilterUpdating };
-export { subscribeToUsersList };
+export { updateUsersState };
