@@ -2,44 +2,77 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 import { Link } from 'react-router-dom';
-
-import translate from '../../../i18n';
-import UsersFilter from '../containers/UsersFilter';
-import UsersList from '../containers/UsersList';
-import { Button, PageLayout } from '../../common/components/web';
+import { compose } from 'react-apollo';
 
 import settings from '../../../../../../settings';
+import translate from '../../../i18n';
+import UsersFilterView from '../components/UsersFilterView';
+import { Button, PageLayout } from '../../common/components/web';
+import UsersListView from '../components/UsersListView';
+import withSubscription from './withSubscription';
+import {
+  withFilterUpdating,
+  withOrderByUpdating,
+  withUsers,
+  withUsersDeleting,
+  withUsersState,
+  updateUsersState
+} from './UserOperations';
 
-const Users = ({ t }) => {
-  const renderMetaData = () => (
-    <Helmet
-      title={`${settings.app.name} - ${t('users.title')}`}
-      meta={[
-        {
-          name: 'description',
-          content: `${settings.app.name} - ${t('users.meta')}`
-        }
-      ]}
-    />
-  );
+class Users extends React.Component {
+  constructor(props) {
+    super(props);
+  }
 
-  return (
-    <PageLayout>
-      {renderMetaData()}
-      <h2>{t('users.list.title')}</h2>
-      <Link to="/users/new">
-        <Button color="primary">{t('users.btn.add')}</Button>
-      </Link>
-      <hr />
-      <UsersFilter />
-      <hr />
-      <UsersList />
-    </PageLayout>
-  );
-};
+  componentDidUpdate() {
+    const { usersUpdated, updateQuery } = this.props;
+    if (usersUpdated) {
+      updateUsersState(usersUpdated, updateQuery);
+    }
+  }
+
+  renderMetaData() {
+    return (
+      <Helmet
+        title={`${settings.app.name} - ${this.props.t('users.title')}`}
+        meta={[
+          {
+            name: 'description',
+            content: `${settings.app.name} - ${this.props.t('users.meta')}`
+          }
+        ]}
+      />
+    );
+  }
+
+  render() {
+    return (
+      <PageLayout>
+        {this.renderMetaData()}
+        <h2>{this.props.t('users.list.title')}</h2>
+        <Link to="/users/new">
+          <Button color="primary">{this.props.t('users.btn.add')}</Button>
+        </Link>
+        <hr />
+        <UsersFilterView {...this.props} />
+        <hr />
+        <UsersListView {...this.props} />
+      </PageLayout>
+    );
+  }
+}
 
 Users.propTypes = {
+  usersUpdated: PropTypes.object,
+  updateQuery: PropTypes.func,
   t: PropTypes.func
 };
 
-export default translate('user')(Users);
+export default compose(
+  withUsersState,
+  withUsers,
+  withUsersDeleting,
+  withOrderByUpdating,
+  withFilterUpdating,
+  withSubscription
+)(translate('user')(Users));
