@@ -1,8 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
+import { connect } from 'formik';
 
-export default class FieldAdapter extends Component {
+class FieldAdapter extends Component {
   static propTypes = {
+    formik: PropTypes.object.isRequired,
     component: PropTypes.func,
     onChangeText: PropTypes.func,
     onChange: PropTypes.func,
@@ -15,15 +17,9 @@ export default class FieldAdapter extends Component {
     disabled: PropTypes.bool
   };
 
-  static contextTypes = {
-    formik: PropTypes.object
-  };
-
-  constructor(props, context) {
-    super(props, context);
-    if (!context.formik) {
-      throw new Error('Field must be inside a component decorated with formik()');
-    }
+  constructor(props) {
+    super(props);
+    this.props = props;
   }
 
   onChange = e => {
@@ -31,13 +27,12 @@ export default class FieldAdapter extends Component {
     if (onChange) {
       onChange(e.target.value, e);
     } else {
-      this.context.formik.handleChange(e);
+      this.props.formik.handleChange(e);
     }
   };
 
   onBlur = e => {
-    const { onBlur, name } = this.props;
-    const { formik } = this.context;
+    const { formik, onBlur, name } = this.props;
     if (onBlur) {
       onBlur(e);
     } else {
@@ -50,19 +45,18 @@ export default class FieldAdapter extends Component {
   };
 
   onChangeText = value => {
-    const { onChangeText, onChange, name } = this.props;
+    const { formik, onChangeText, onChange, name } = this.props;
     if (onChange && !onChangeText) {
       onChange(value);
     } else if (onChangeText) {
       onChangeText(value);
     } else {
-      this.context.formik.setFieldValue(name, value);
+      formik.setFieldValue(name, value);
     }
   };
 
   render() {
-    const { formik } = this.context;
-    const { component, name, defaultValue, defaultChecked, disabled } = this.props;
+    const { formik, component, name, defaultValue, defaultChecked, disabled } = this.props;
     let { value, checked } = this.props;
     value = value || '';
     checked = checked || false;
@@ -93,3 +87,5 @@ export default class FieldAdapter extends Component {
     });
   }
 }
+
+export default connect(FieldAdapter);
