@@ -1,14 +1,11 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
 
-import { SchemaLink } from 'apollo-link-schema';
 import WebSocket from 'ws';
 
 import createApolloClient from '../../../common/createApolloClient';
 import '../../knexdata';
 import knex from '../sql/connector';
-import schema from '../api/schema';
-import modules from '../modules';
 
 chai.use(chaiHttp);
 chai.should();
@@ -22,19 +19,18 @@ before(async () => {
   await knex.migrate.latest();
   await knex.seed.run();
 
-  server = require('../server').default;
+  server = await require('../server').default;
 
-  const context = await modules.createContext();
   global.WebSocket = WebSocket;
   apollo = createApolloClient({
-    apiUrl: `http://localhost:${process.env['PORT']}/graphql`,
-    createNetLink: () => new SchemaLink({ schema, context })
+    apiUrl: `http://localhost:${process.env['PORT']}/graphql`
   });
 });
 
 after(() => {
   if (server) {
     server.close();
+    delete global.__TEST_SESSION__;
   }
 });
 
