@@ -1,3 +1,4 @@
 declare var __SERVER_PORT__: number;
 declare var __API_URL__: string;
-
+declare var __FRONTEND_BUILD_DIR__: string;
+declare var __SSR__: boolean;
