@@ -4,20 +4,119 @@ const addModule = require('./commands/addModule');
 const deleteModule = require('./commands/deleteModule');
 const updateSchema = require('./commands/updateSchema');
 
-module.exports = (action, args, options, logger) => {
-  const module = args.module;
-  let location = 'both';
-  if (args.location) {
-    location = args.location;
+String.prototype.toCamelCase = function() {
+  return this.replace(/^([A-Z])|\s(\w)/g, function(match, p1, p2) {
+    if (p2) return p2.toUpperCase();
+    return p1.toLowerCase();
+  });
+};
+
+String.prototype.capitalize = function() {
+  return this.charAt(0).toUpperCase() + this.slice(1);
+};
+
+function copyFiles(logger, templatePath, module, location) {
+  logger.info(`Copying ${location} files…`);
+
+  // create new module directory
+  const mkdir = shell.mkdir(`${__dirname}/../../packages/${location}/src/modules/${module}`);
+
+  // continue only if directory does not jet exist
+  if (mkdir.code === 0) {
+    const destinationPath = `${__dirname}/../../packages/${location}/src/modules/${module}`;
+    shell.cp('-R', `${templatePath}/${location}/*`, destinationPath);
+
+    logger.info(`✔ The ${location} files have been copied!`);
+
+    // change to destination directory
+    shell.cd(destinationPath);
+
+    // rename files
+    shell.ls('-Rl', '.').forEach(entry => {
+      if (entry.isFile()) {
+        const moduleFile = entry.name.replace('Module', module.capitalize());
+        shell.mv(entry.name, moduleFile);
+      }
+    });
+
+    // replace module names
+    shell.ls('-Rl', '.').forEach(entry => {
+      if (entry.isFile()) {
+        shell.sed('-i', /\$module\$/g, module, entry.name);
+        shell.sed('-i', /\$Module\$/g, module.toCamelCase().capitalize(), entry.name);
+      }
+    });
+
+    // get index file path
+    const modulesPath = `${__dirname}/../../packages/${location}/src/modules/`;
+    const indexFullFileName = fs.readdirSync(modulesPath).find(name => name.search(/index/) >= 0);
+    const indexPath = modulesPath + indexFullFileName;
+    let indexContent;
+
+    try {
+      // prepend import module
+      indexContent = `import ${module} from './${module}';\n` + fs.readFileSync(indexPath);
+    } catch (e) {
+      logger.error(`Failed to read ${indexPath} file`);
+      process.exit();
+    }
+
+    // extract Feature modules
+    const featureRegExp = /Feature\(([^()]+)\)/g;
+    const [, featureModules] = featureRegExp.exec(indexContent) || ['', ''];
+
+    // add module to Feature connector
+    shell
+      .ShellString(indexContent.replace(RegExp(featureRegExp, 'g'), `Feature(${module}, ${featureModules})`))
+      .to(indexPath);
+
+    logger.info(`✔ Module for ${location} successfully created!`);
   }
-  let tablePrefix = '';
-  if (args.tablePrefix) {
-    tablePrefix = args.tablePrefix;
-  }
-  console.log(tablePrefix);
-  let templatePath = `${__dirname}/../templates/module`;
-  if (action === 'addcrud') {
-    templatePath = `${__dirname}/../templates/crud`;
+}
+
+function deleteFiles(logger, templatePath, module, location) {
+  logger.info(`Deleting ${location} files…`);
+
+  const modulePath = `${__dirname}/../../packages/${location}/src/modules/${module}`;
+
+  if (fs.existsSync(modulePath)) {
+    // delete module directory
+    shell.rm('-rf', modulePath);
+
+    // path to modules index file
+    const modulesPath = `${__dirname}/../../packages/${location}/src/modules/`;
+
+    // get index file path
+    const indexFullFileName = fs.readdirSync(modulesPath).find(name => name.search(/index/) >= 0);
+    const indexPath = modulesPath + indexFullFileName;
+    let indexContent;
+
+    try {
+      indexContent = fs.readFileSync(indexPath);
+    } catch (e) {
+      logger.error(`Failed to read ${indexPath} file`);
+      process.exit();
+    }
+
+    // extract Feature modules
+    const featureRegExp = /Feature\(([^()]+)\)/g;
+    const [, featureModules] = featureRegExp.exec(indexContent) || ['', ''];
+    const featureModulesWithoutDeleted = featureModules
+      .split(',')
+      .filter(featureModule => featureModule.trim() !== module);
+
+    const contentWithoutDeletedModule = indexContent
+      .toString()
+      // replace features modules on features without deleted module
+      .replace(featureRegExp, `Feature(${featureModulesWithoutDeleted.toString().trim()})`)
+      // remove import module
+      .replace(RegExp(`import ${module} from './${module}';\n`, 'g'), '');
+
+    fs.writeFileSync(indexPath, contentWithoutDeletedModule);
+
+    logger.info(`✔ Module for ${location} successfully deleted!`);
+  } else {
+    logger.info(`✔ Module ${location} location for ${modulePath} wasn't found!`);
   }
 
   if (!fs.existsSync(templatePath)) {
@@ -47,4 +146,4 @@ module.exports = (action, args, options, logger) => {
   if (action === 'updateschema') {
     updateSchema(logger, module);
   }
-};
+}
